src/requirements.txt
numpy==2.1.2
matplotlib==3.9.2
python-dotenv==1.0.1 
